refactor moon tool
