+++ crates/moon/src/cli/tool.rs
@@ -16,7 +16,9 @@
 //
 // For inquiries, you can contact us via e-mail at jichuruanjian@idea.edu.cn.
 
-use std::{io::BufRead, path::PathBuf, process::Stdio};
+mod format_and_diff;
+
+use format_and_diff::*;
 
 #[derive(Debug, clap::Parser)]
 pub struct ToolSubcommand {
@@ -29,77 +31,8 @@ pub enum ToolSubcommands {
     FormatAndDiff(FormatAndDiffSubcommand),
 }
 
-#[derive(Debug, clap::Parser)]
-pub struct FormatAndWriteSubcommand {
-    #[clap(long)]
-    old: PathBuf,
-    #[clap(long)]
-    new: PathBuf,
-}
-
-#[derive(Debug, clap::Parser)]
-pub struct FormatAndDiffSubcommand {
-    #[clap(long)]
-    old: PathBuf,
-    #[clap(long)]
-    new: PathBuf,
-}
-
 pub fn run_tool(cmd: ToolSubcommand) -> anyhow::Result<i32> {
     match cmd.subcommand {
         ToolSubcommands::FormatAndDiff(subcmd) => run_format_and_diff(subcmd),
-    }
-}
-
-fn run_format_and_diff(cmd: FormatAndDiffSubcommand) -> anyhow::Result<i32> {
-    let mut execution = std::process::Command::new("moonfmt")
-        .arg("-exit-code")
-        .arg(cmd.old.to_str().unwrap())
-        .arg("-o")
-        .arg(cmd.new.to_str().unwrap())
-        .stdin(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .spawn()?;
-
-    let x = execution.wait()?;
-    let exit_code = x.code().unwrap_or(1);
-    if exit_code == 0 {
-        return Ok(0);
-    }
-    let mut execution = std::process::Command::new("git")
-        .args([
-            "--no-pager",
-            "diff",
-            "--color=always",
-            "--no-index",
-            cmd.old.to_str().unwrap(),
-            cmd.new.to_str().unwrap(),
-        ])
-        .stdout(Stdio::piped())
-        .stderr(Stdio::inherit())
-        .spawn()?;
-    let child_stdout = execution.stdout.take().unwrap();
-    let mut buf = String::new();
-    let mut bufread = std::io::BufReader::new(child_stdout);
-    while let Ok(n) = bufread.read_line(&mut buf) {
-        if n > 0 {
-            print!("{}", buf);
-            buf.clear()
-        } else {
-            break;
-        }
-    }
-    let status = execution.wait()?;
-    match status.code() {
-        Some(0) => Ok(0),
-        _ => {
-            eprintln!(
-                "failed to execute `git --no-pager diff --color=always --no-index {} {}`",
-                &cmd.old.to_str().unwrap(),
-                &cmd.new.to_str().unwrap()
-            );
-            Ok(1)
-        }
     }
 }
